cast-operacion agregado, sentencias de control, if, while, switch, sus esqueletos fueron agregados

+++ proyect-ide-compilador/src/app/logic/instructions/operations/nodo-operation.ts
@@ -57,7 +57,7 @@ export class NodoOperation {
       case TypeDato.INTEGER:
         this.result = `${this.dato.numero}`;
         break;
-      case TypeDato.DOUBLE:
+      case TypeDato.FLOAT:
         this.result = `${this.dato.numero}`;
         break;
       case TypeDato.BOOLEAN:
